Cleaning up the Embassy's Bureaucracy
I used the wonderful features of "Lambda expressions" and interfaces to
eliminate excess lines of code. Now, rather than each method containing
all the lines of code which execute the SQL mechanics, those are
contained in two methods, query() and update(), which simply take in
whatever query is to be run. The query() method also accepts a "Process"
implementation, which takes care of the particulars of what needs to be
done with the ResultSet of the query.